Organized with hexagonal structure and cleaned the code
