+++ src/app/modules/users/controllers/users.controller.ts
@@ -8,9 +8,9 @@ import {
   Delete,
 } from '@nestjs/common';
 import { UsersService } from '../services/users.service';
-import { CreateUserDto } from '../dto/create-user.dto';
+import { CreateUserDto } from '../dtos/create-user.dto';
 import { ApiTags } from '@nestjs/swagger';
-import { JwtAuthGuard } from 'src/guards/jwt-auth.guard';
+import { JwtAuthGuard } from '../../../../guards/jwt-auth.guard';
 
 @ApiTags('Users')
 @Controller('users')
