--- src/authentication/services/authentication.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AuthenticationService } from './authentication.service';
-
-describe('AuthenticationService', () => {
-  let service: AuthenticationService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AuthenticationService],
-    }).compile();
-
-    service = module.get<AuthenticationService>(AuthenticationService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
